move image signing tests to e2e
